+++ src/main/java/com/mercadolibre/dto/remedy/RemediesRequest.java
@@ -1,6 +1,5 @@
 package com.mercadolibre.dto.remedy;
 
-import com.mercadolibre.px.toolkit.dto.user_agent.UserAgent;
 import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
@@ -13,7 +12,6 @@ public class RemediesRequest {
 
   @Setter private long riskExcecutionId;
   @Setter private String statusDetail;
-  @Setter private UserAgent userAgent;
   @Setter private String siteId;
   @Setter private String userId;
   @Setter private boolean oneTap;